update figma code thread handlers structure
