+++ app/lib/main/index.tsx
@@ -169,9 +169,6 @@ export default function App(props: { platform: TargetPlatform }) {
 
   const updateFocusedScreen = (screen: WorkScreen) => {
     // notify code.ts that app mode has set.
-    console.log(
-      `sending back thread about changed screen (user interest) data - "${screen}"`
-    );
     parent.postMessage(
       {
         pluginMessage: {
@@ -180,6 +177,9 @@ export default function App(props: { platform: TargetPlatform }) {
         },
       },
       "*"
+    );
+    console.log(
+      `sending back thread about changed screen (user interest) data - "${screen}"`
     );
   };
