+++ figma/src/code-thread/core-init.ts
@@ -1,0 +1,10 @@
+// import this file
+
+import { __register__ as __reguster_default_handler__ } from "./default-handler";
+import { __register__ as __register_resize_screen__ } from "./resize-screen";
+import { __register__ as __register_user_interest__ } from "./user-interest";
+//
+__reguster_default_handler__();
+__register_resize_screen__();
+__register_user_interest__();
+//
