+++ figma/src/event-handlers/README.md
@@ -1,0 +1,1 @@
+Look at init.ts
