+++ figma/src/event-handlers/index.ts
@@ -1,0 +1,8 @@
+// DO NOT IMPORT ME UNLESS HOST IS CODE.TS
+
+///
+/// simply importing this will register utility handlers.
+///
+
+// index wrapper for importing init
+import "./init";
