+++ figma/src/event-handlers/init.ts
@@ -1,0 +1,11 @@
+import { __register__ as __register_create_button__ } from "./create-button";
+import { __register__ as __register_create_icon__ } from "./create-icon";
+import { __register__ as __register_hide_all__ } from "./hide-all";
+import { __register__ as __register_randomize_selection__ } from "./randomize-selection";
+import { __register__ as __register_replace_font__ } from "./replace-font";
+//
+__register_create_button__();
+__register_create_icon__();
+__register_hide_all__();
+__register_randomize_selection__();
+__register_replace_font__();
