+++ figma/src/event-handlers/randomize-selection.ts
@@ -1,0 +1,8 @@
+import { randimizeText } from "../tool-box/manipulate";
+import { addEventHandler } from "../code-thread";
+
+export function __register__() {
+  addEventHandler("randomize-selection", () => {
+    randimizeText();
+  });
+}
